requirements.txt
datasets==4.0.0
transformers==4.56.1
sentence-transformers==5.1.0
faiss-cpu==1.12.0
torch==2.8.0+cu126
gTTS==2.5.4
gradio==5.44.1